pyproject.toml
[project]
name = "livelaucher"
version = "0.1.0"
description = "Sequential service launcher with port connectivity check"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.1",
]

[project.scripts]
livelaucher = "main:main"
